requirements.txt
asgiref @ file:///tmp/build/80754af9/asgiref_1594338739818/work
astroid==2.4.2
beautifulsoup4==4.9.1
bokeh==2.2.0
cairocffi==1.1.0
CairoSVG==2.4.2
certifi==2020.6.20
cffi==1.14.1
colorama==0.4.3
cssselect2==0.3.0
defusedxml==0.6.0
diff-match-patch==20200713
Django==3.0.8
django-bootstrap3==14.1.0
django-bootstrap4==2.2.0
django-easy-pdf==0.1.1
django-import-export==2.3.0
django-ipyfield==0.1.6
django-mathfilters==1.0.0
django-wkhtmltopdf==3.3.0
-e git+https://github.com/chrisglass/django-xhtml2pdf.git@d3d7dc527788953adf02a6d99e37dee225bd8c51#egg=django_xhtml2pdf
et-xmlfile==1.0.1
html5lib==1.1
inflect==4.1.0
IPy==1.0
isort==4.3.21
jdcal==1.4.1
Jinja2==2.11.2
lazy-object-proxy==1.4.3
MarkupPy==1.14
MarkupSafe==1.1.1
mccabe==0.6.1
numpy==1.19.1
odfpy==1.4.1
openpyxl==3.0.4
packaging==20.4
pdfkit==0.6.1
Pillow==7.2.0
pycparser==2.20
pylint==2.5.3
pyparsing==2.4.7
PyPDF2==1.26.0
Pyphen==0.9.5
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3.1
reportlab==3.5.46
schedule==0.6.0
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
style==1.1.0
tablib==2.0.0
tinycss2==1.0.2
toml==0.10.1
tornado==6.0.4
typing-extensions==3.7.4.3
update==0.0.1
WeasyPrint==51
webencodings==0.5.1
wincertstore==0.2
wkhtmltopdf==0.2
wrapt==1.12.1
xhtml2pdf==0.2.4
xlrd==1.2.0
xlwt==1.3.0
